+++ monetdb5/mal/mal_import.c
@@ -154,7 +154,7 @@ malInclude(Client c, str name, int listing)
 	str p;
 
 	bstream *oldfdin = c->fdin;
-	int oldyycur = c->yycur;
+	size_t oldyycur = c->yycur;
 	int oldlisting = c->listing;
 	enum clientmode oldmode = c->mode;
 	int oldblkmode = c->blkmode;
